update: add swagger

+++ src/utils/save-to-cloudinary.js
@@ -12,6 +12,7 @@ cloudinary.v2.config({
 });
 
 export const saveFileToCloudinary = async (file) => {
+  if (!file) return;
   const response = await cloudinary.v2.uploader.upload(file.path);
   await fs.unlink(file.path);
   return response.secure_url;
